@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :destination_display, :class => Chouette::DestinationDisplay do
     sequence(:name) { |n| "DestinationDisplay name #{n}" }
     sequence(:front_text) { |n| "DestinationDisplay front_text #{n}" }
