@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   
   factory :journey_pattern_common, :class => Chouette::JourneyPattern do
     sequence(:name) { |n| "jp name #{n}" }
