@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
 
   factory :stop_point, :class => Chouette::StopPoint do
     sequence(:objectid) { |n| "test:StopPoint:#{n}" }
