@@ -12,7 +12,7 @@ feature 'Sign out', :devise do
   #   When I sign out
   #   Then I see a signed out message
   scenario 'user signs out successfully' do
-    user = FactoryGirl.create(:user)
+    user = FactoryBot.create(:user)
     user.confirm
     signin(user.email, user.password)
     expect(page).to have_content I18n.t 'devise.sessions.signed_in'
