@@ -18,7 +18,7 @@ feature 'User edit', :devise do
   #   When I change my email address
   #   Then I see an account updated message
   scenario 'user changes email address' do
-    user = FactoryGirl.create(:user)
+    user = FactoryBot.create(:user)
     user.confirm
     login_as(user, :scope => :user)
     visit edit_user_registration_path(user)
@@ -34,9 +34,9 @@ feature 'User edit', :devise do
   #   When I try to edit another user's profile
   #   Then I see my own 'edit profile' page
   scenario "user cannot cannot edit another user's profile", :me do
-    me = FactoryGirl.create(:user)
+    me = FactoryBot.create(:user)
     me.confirm
-    other = FactoryGirl.create(:user, email: 'other@example.com')
+    other = FactoryBot.create(:user, email: 'other@example.com')
     other.confirm
     login_as(me, :scope => :user)
     visit edit_user_registration_path(other)
