@@ -7,23 +7,23 @@ describe Chouette::StopArea do
 
     it "should return referential projection " do
       subject.referential.projection_type='27572'
-      subject.projection.should == subject.referential.projection_type
+      expect(subject.projection).to eq(subject.referential.projection_type)
     end
 
     it "should return projection coordinates when referential has projection" do
       subject.latitude = 45
       subject.longitude = 0
       subject.referential.projection_type='27572'
-      subject.projection_x.should_not be_nil
-      subject.projection_y.should_not be_nil
+      expect(subject.projection_x).not_to be_nil
+      expect(subject.projection_y).not_to be_nil
     end
 
     it "should return nil projection coordinates when referential has no projection" do
       subject.latitude = 45
       subject.longitude = 0
       subject.referential.projection_type=nil
-      subject.projection_x.should be_nil
-      subject.projection_y.should be_nil
+      expect(subject.projection_x).to be_nil
+      expect(subject.projection_y).to be_nil
     end
 
 end
@@ -35,23 +35,23 @@ describe Chouette::AccessPoint do
 
     it "should return referential projection " do
       subject.referential.projection_type='27572'
-      subject.projection.should == subject.referential.projection_type
+      expect(subject.projection).to eq(subject.referential.projection_type)
     end
 
     it "should return projection coordinates when referential has projection" do
       subject.latitude = 45
       subject.longitude = 0
       subject.referential.projection_type='27572'
-      subject.projection_x.should_not be_nil
-      subject.projection_y.should_not be_nil
+      expect(subject.projection_x).not_to be_nil
+      expect(subject.projection_y).not_to be_nil
     end
 
     it "should return nil projection coordinates when referential has no projection" do
       subject.latitude = 45
       subject.longitude = 0
       subject.referential.projection_type=nil
-      subject.projection_x.should be_nil
-      subject.projection_y.should be_nil
+      expect(subject.projection_x).to be_nil
+      expect(subject.projection_y).to be_nil
     end
 
 end
