@@ -7,6 +7,6 @@ describe Organisation do
 
   it "create a rule_parameter_set" do
     organisation = create(:organisation)
-    organisation.rule_parameter_sets.size.should == 1
+    expect(organisation.rule_parameter_sets.size).to eq(1)
   end
 end
