@@ -2,7 +2,7 @@ require 'spec_helper'
 
 describe ExportTask, :type => :model do
 
-  it { should_not validate_presence_of(:start_date) }
-  it { should_not validate_presence_of(:end_date) }
+  it { is_expected.not_to validate_presence_of(:start_date) }
+  it { is_expected.not_to validate_presence_of(:end_date) }
 
 end
