@@ -122,7 +122,7 @@ describe VehicleJourneyImport, :type => :model do
 
     it "should return false when stop points in file are not the same in the route" do
       vehicle_journey_import = VehicleJourneyImport.new(:route => other_route, :file => valid_file)
-      expect { vehicle_journey_import.load_imported_vehicle_journeys }.to raise_exception
+      expect { vehicle_journey_import.load_imported_vehicle_journeys }.to raise_exception(RuntimeError)
     end
 
     # it "should return errors when vehicle journeys in file are invalid" do            
@@ -133,7 +133,7 @@ describe VehicleJourneyImport, :type => :model do
 
     it "should return errors when vehicle journey at stops in file are invalid" do           
       vehicle_journey_import = VehicleJourneyImport.new(:route => route, :file => invalid_file_on_vjas)
-      expect { vehicle_journey_import.load_imported_vehicle_journeys }.to raise_exception
+      expect { vehicle_journey_import.load_imported_vehicle_journeys }.to raise_exception(ArgumentError)
     end
 
     it "should return errors when vehicle journey at stops are not in ascendant order" do    
